dian

+++ database/migrations/2026_05_23_081344_add_geosite_to_informasi_table.php
@@ -22,7 +22,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('informasi', function (Blueprint $table) {
-            $table->dropColumn('geosite');
+            if (Schema::hasColumn('informasi', 'geosite') && \DB::getDriverName() === 'mysql') {
+                $table->dropColumn('geosite');
+            }
         });
     }
 };
